Fix compilation warnings
- Remove spurious #pragma message in MarchingCubes.cpp and PoissonRecon.cpp
- Add parentheses around assignment in MultiGridOctreeData.IsoSurface.inl
- Add return statement for non-void function
- Replace deprecated vsprintf with vsnprintf in test_mocks.h
- Cast RAND_MAX to float in test files to fix implicit conversion warning

+++ tests/test_reconstruction.h
@@ -72,8 +72,8 @@ void test_reconstruct_sphere() {
     std::vector<std::pair<Point3D<float>, Point3D<float>>> points(numPoints);
     
     for (int i = 0; i < numPoints; ++i) {
-        float u = (float)rand() / RAND_MAX;
-        float v = (float)rand() / RAND_MAX;
+        float u = (float)rand() / (float)RAND_MAX;
+        float v = (float)rand() / (float)RAND_MAX;
         float theta = 2.0f * (float)M_PI * u;
         float phi = acos(2.0f * v - 1.0f);
